requirements.txt
amqp==2.1.1
billiard==3.5.0.2
celery==4.0.0
click==6.6
coverage==4.3.4
Django==1.10
django-appconf==1.0.2
django-jsonfield==1.0.1
django-registration-redux==1.4
Flask==0.11.1
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
kombu==4.0.0
MarkupSafe==0.23
matlabengineforpython===R2016b
Pillow==3.3.0
pytz==2016.7
requests==2.11.1
vine==1.1.3
Werkzeug==0.11.11
